Fix #1361 hoist core agent profile defaults
