+++ src/pollypm/accounts.py
@@ -11,7 +11,7 @@ from pathlib import Path
 
 import typer
 
-from pollypm.plugins_builtin.core_agent_profiles.profiles import heartbeat_prompt, polly_prompt
+from pollypm.agent_profiles.defaults import heartbeat_prompt, polly_prompt
 from pollypm.config import load_config, write_config
 from pollypm.models import AccountConfig, PollyPMConfig, ProviderKind
 from pollypm.onboarding import (
